Addressed Lu code review, started on NoPreviousVersionAdapter
